v2.0.0

+++ PCAB_Debugger2_GUI/cntPorts.xaml.cs
@@ -1,17 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace PCAB_Debugger2_GUI
 {
@@ -62,7 +52,7 @@ namespace PCAB_Debugger2_GUI
             {
                 for (int i = 0; i < value.Count; i++)
                 {
-                    SetDSA((uint)(i + 1), (int)value[i]);
+                    SetDSA((uint)(i + 1), value[i]);
                 }
             }
         }
@@ -137,7 +127,7 @@ namespace PCAB_Debugger2_GUI
             {
                 for (int i = 0; i < value.Count; i++)
                 {
-                    SetDPS((uint)(i + 1), (int)value[i]);
+                    SetDPS((uint)(i + 1), value[i]);
                 }
             }
         }
